Refactor single stat colors UI into generic components
